requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
httpx
streamlit
dateparser
google-api-python-client
google-auth
google-auth-oauthlib
pytest
pytz
requests
pyttsx3
SpeechRecognition
pyaudio
vonage
